Remove text on 800-227
This text was (I think) added based on my comment that it would be good to remind readers that FIPS 203 states:

"For general definitions and properties of KEMs, including requirements for the secure use of KEMs in applications, see SP 800-227"

The added text talks about "general guidance" and is in my view making things worse. Instead of spending time discussing exactly which requirements in 800-227 that apply to TLS 1.3, I suggest just removing the newly added text on 800-227 and publish. The draft does not have to duplicate text from NIST specifications.